+++ CHANGELOG.md
@@ -28,6 +28,8 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Fixed
 
+- Close release-created branch-stack validation PRs after successful publication instead of leaving obsolete conflicted PRs open.
+
 ### Security
 
 ## [3.1.0] - 2026-09-10